requirements.txt
streamlit==1.33.0
langchain==0.1.13
python-docx==1.1.0
tqdm==4.66.2
pydantic==1.10.13
python-dotenv
transformers 
torch
python-docx
pypdf
langgraph
langchain-community 
langchain-core 
langchain_ollama
langchain-qdrant qdrant-client
